Revise current statuses
Given the feedback of the team, I think we should add a `Reopened` status and change `Feedback` to `Waiting`. `Reopened` draws attention to the fact that a previously completed ticket was reopened and that it's not `New`. `Waiting` is a more direct way to say "I'm blocked" and that you cannot continue work without someone else getting involved or helping.